+++ generators/app/index.js
@@ -108,6 +108,7 @@ const obj = {
       type: 'list',
       choices: [
         'eureka',
+        'nacos',
         // 'zookeeper',
         'none'
       ],
@@ -241,6 +242,25 @@ const obj = {
       }
     }
   },
+  docker: {
+    prompting: {
+      type: 'list',
+      choices: [
+        { key: 'Jib', display: 'Jib (推荐使用)' },
+        'Dockerfile',
+        'dockerfile-maven-plugin'
+      ],
+      message: '请选择镜像构建方式'
+    },
+    option: { desc: '镜像构建方式', type: String, default: 'Dockerfile' },
+    child: {
+      jdk: {
+        prompting: { type: 'input', default: 'openjdk:8', message: '请输入 JDK ' },
+        option: { desc: 'Dockerfile FROM', type: String, default: 'openjdk:8' }
+      }
+    }
+
+  },
   prometheus: {
     prompting: {
       type: 'confirm',
@@ -248,6 +268,14 @@ const obj = {
       default: false
     },
     option: { desc: 'Prometheus', type: Boolean, default: false }
+  },
+  mongodb: {
+    prompting: {
+      type: 'confirm',
+      message: '是否整合 mongodb（默认no）',
+      default: false
+    },
+    option: { desc: 'mongodb', type: Boolean, default: false }
   },
   quartz: {
     prompting: {
@@ -310,7 +338,7 @@ module.exports = require('yo-power-generator').getGenerator(obj, {
       props.conditions[props.authentication] = true;
     }
 
-    if (props.security !== 'none') {
+    if (props.security && props.security !== 'none') {
       props.conditions[props.security] = true;
     }
 
@@ -329,6 +357,18 @@ module.exports = require('yo-power-generator').getGenerator(obj, {
     if (!props.log) {
       props.log = 'logback'
     }
+
+    if (props.docker) {
+      if (props.docker === 'Dockerfile') {
+        props.conditions[props.docker] = true;
+      } else {
+        props.conditions['dockerScripts'] = true;
+        if (props.docker === 'dockerfile-maven-plugin') {
+          props.conditions['dockerMvn'] = true;
+        }
+      }
+    }
+
     props.conditions[props.log] = true;
 
     props.openfeign = props.discovery === 'eureka';