+++ generators/app/templates/${artifactId}-provider/src/main/java/${basePath}/controller/if_demo_nacos.NacosDemoController.tmpl.java
@@ -1,0 +1,26 @@
+package ${basePackage}.controller;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.context.config.annotation.RefreshScope;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/config")
+@RefreshScope
+@Payload
+public class NacosDemoController {
+
+<%
+    let tmpl = `    @Value("\$\{useLocalCache:false\}")`;
+    print(tmpl)
+%>
+    private boolean useLocalCache;
+
+
+    @GetMapping("/get")
+    public boolean get() {
+        return useLocalCache;
+    }
+}