+++ generators/app/templates/${artifactId}-provider/src/main/java/${basePath}/domain/dto/if_demo_mongodb.DocumentDemoDTO.tmpl.java
@@ -1,0 +1,13 @@
+package ${basePackage}.domain.dto;
+
+import lombok.Data;
+import lombok.experimental.Accessors;
+
+import javax.validation.constraints.NotNull;
+
+@Data
+@Accessors(chain = true)
+public class DocumentDemoDTO {
+    @NotNull
+    private String username;
+}