+++ generators/app/templates/${artifactId}-provider/src/main/java/${basePath}/domain/if_demo_mongodb.DocumentDemo.tmpl.java
@@ -1,0 +1,19 @@
+package ${basePackage}.domain;
+
+import lombok.Data;
+import lombok.experimental.Accessors;
+import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.mapping.Document;
+import org.springframework.data.mongodb.core.mapping.Field;
+
+import java.io.Serializable;
+
+@Data
+@Accessors(chain = true)
+@Document(collection = "user")
+public class DocumentDemo {
+    @Id
+    private String id;
+    @Field("username")
+    private String username;
+}