+++ generators/test/1.js
@@ -1,0 +1,9 @@
+const _ = require('lodash')
+const fs = require('fs');
+const path = require('path');
+
+const content = fs.readFileSync(path.join(__dirname, '/1.tmpl'), { encoding: 'utf-8' });
+console.log(_.template(content, {})({
+  artifactId: 'test',
+  docker: 'Dockerfile'
+}))