+++ CHANGELOG.md
@@ -2,6 +2,38 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.19.0](https://github.com/deepexi/generator-deepexi-spring-cloud/compare/v1.18.0...v1.19.0) (2020-04-14)
+
+
+### Bug Fixes
+
+* error content on option dockerfile ([981c788](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/981c788))
+* missing import of @NotNull package ([5214d04](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/5214d04))
+* should only cover bean MessageConverter ([222e9ba](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/222e9ba))
+
+
+### Features
+
+* provide '.gitlab-ci.yaml' by default ([#21](https://github.com/deepexi/generator-deepexi-spring-cloud/issues/21)) ([9114f08](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/9114f08))
+* support integrate with mongodb ([#17](https://github.com/deepexi/generator-deepexi-spring-cloud/issues/17)) ([c420587](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/c420587))
+
+
+### Tests
+
+* add test files ([8fcde01](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/8fcde01))
+
+
+
+## [1.18.0](https://github.com/deepexi/generator-deepexi-spring-cloud/compare/v1.17.0...v1.18.0) (2020-03-02)
+
+
+### Features
+
+* enhancements for docker ([#18](https://github.com/deepexi/generator-deepexi-spring-cloud/issues/18)) ([19c6bb8](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/19c6bb8))
+* support integration with prometheus([#11](https://github.com/deepexi/generator-deepexi-spring-cloud/issues/11)) ([db297a5](https://github.com/deepexi/generator-deepexi-spring-cloud/commit/db297a5))
+
+
+
 ## [1.17.0](https://github.com/deepexi/generator-deepexi-spring-cloud/compare/v1.16.0...v1.17.0) (2019-11-09)